requirements.txt
# Smart Peak Hour Predictor - Complete Requirements
# Compatible with Python 3.10 only
# Do not upgrade pytorch-forecasting or pytorch-lightning for Temple Fusion compatibility

# Core ML Framework
torch>=2.0.0,<2.1.0
pytorch-forecasting>=1.0.0,<1.1.0
pytorch-lightning>=2.0.0,<2.1.0
torchmetrics==1.5.2

# Data Processing & Analysis
pandas>=1.5.0,<2.0.0
numpy>=1.23.0,<1.24.0
scikit-learn>=1.0.0,<1.1.0
scipy==1.15.3
statsmodels==0.14.4

# Visualization & UI
matplotlib==3.8.4
seaborn==0.12.2
streamlit>=1.24.0,<1.25.0  # Latest stable version
altair==5.5.0
pydeck==0.9.1
plotly>=5.13.0,<5.14.0  # Required for interactive visualizations

# Logging & Monitoring
tensorboard==2.19.0
tensorboard-data-server==0.7.2
logging>=0.5.1.2  # Python's built-in logging module

# Utilities & Progress
tqdm==4.66.4
python-dateutil>=2.8.2
pytz>=2023.3
tzdata==2025.2
tzlocal==5.3.1

# HTTP & Networking
requests>=2.28.0,<2.29.0
aiohttp==3.12.12
grpcio==1.73.0
urllib3==2.4.0

# Data Formats & Serialization
protobuf==4.25.8
PyYAML>=6.0,<6.1  # Required for config files
pyarrow==20.0.0

# Development & Build Tools
setuptools>=41.0.0
packaging==23.2
wheel

# Optional Dependencies (for enhanced functionality)
optuna==2.10.1
lightning-utilities==0.14.3

# Weather & Holiday Integration
holidays>=0.16,<0.17  # Required for holiday features
python-dotenv>=1.0.0,<1.1.0  # For environment variables

# Additional ML & Time Series
dvc>=2.0.0,<2.1.0
lightgbm>=3.3.0,<3.4.0
xgboost>=1.7.0,<1.8.0
prophet>=1.1.0,<1.2.0
pmdarima>=2.0.0,<2.1.0
joblib>=1.2.0,<1.3.0

# Database
sqlite3  # Built into Python 3.10

# Note: Some packages may have interdependencies. If you encounter issues,
# try installing in the order listed above.
